создаёт и подключает модуль data
создаёт и подключает модуль для хранения констант
